requirements.txt
Django==1.4.2
South==0.7.6
django-tastypie==0.9.12-alpha
ipython==0.13.1
mimeparse==0.1.3
psycopg2==2.4.5
pyjade==1.6
python-dateutil==1.5
wsgiref==0.1.2
